adding EthType to secure handling ether types
this will also pretty print ethertypes in flow output

+++ core/common/src/main/java/org/onosproject/codec/impl/EncodeCriterionCodec.java
@@ -165,7 +165,7 @@ public final class EncodeCriterionCodec {
         public ObjectNode encodeCriterion(ObjectNode root, Criterion criterion) {
             final EthTypeCriterion ethTypeCriterion =
                     (EthTypeCriterion) criterion;
-            return root.put(CriterionCodec.ETH_TYPE, ethTypeCriterion.ethType());
+            return root.put(CriterionCodec.ETH_TYPE, ethTypeCriterion.ethType().toShort());
         }
     }
